+++ lib/unleash/metrics_reporter.rb
@@ -1,5 +1,4 @@
 require 'unleash/configuration'
-require 'unleash/metrics'
 require 'net/http'
 require 'json'
 require 'time'
@@ -15,52 +14,39 @@ module Unleash
     end
 
     def generate_report
-      now = Time.now
+      metrics = Unleash&.engine&.get_metrics()
+      return nil if metrics.nil? || metrics.empty?
 
-      start = self.last_time
-      stop  = now
-      self.last_time = now
-
-      report = {
-        'appName': Unleash.configuration.app_name,
-        'instanceId': Unleash.configuration.instance_id,
+      {
         'platformName': RUBY_ENGINE,
         'platformVersion': RUBY_VERSION,
-        'yggdrasilVersion': nil,
+        'yggdrasilVersion': "0.13.2",
         'specVersion': Unleash::CLIENT_SPECIFICATION_VERSION,
-        'bucket': {
-          'start': start.iso8601(Unleash::TIME_RESOLUTION),
-          'stop': stop.iso8601(Unleash::TIME_RESOLUTION),
-          'toggles': Unleash.toggle_metrics.features
-        }
+        'appName': Unleash.configuration.app_name,
+        'instanceId': Unleash.configuration.instance_id,
+        'bucket': metrics
       }
-      Unleash.toggle_metrics.reset
-
-      report
     end
 
     def post
       Unleash.logger.debug "post() Report"
 
-      if bucket_empty? && (Time.now - self.last_time < LONGEST_WITHOUT_A_REPORT) # and last time is less then 10 minutes...
+      bucket = self.generate_report
+      if bucket.nil? && (Time.now - self.last_time < LONGEST_WITHOUT_A_REPORT) # and last time is less then 10 minutes...
         Unleash.logger.debug "Report not posted to server, as it would have been empty. (and has been empty for up to 10 min)"
 
         return
       end
 
-      response = Unleash::Util::Http.post(Unleash.configuration.client_metrics_uri, self.generate_report.to_json)
+      response = Unleash::Util::Http.post(Unleash.configuration.client_metrics_uri, bucket.to_json)
 
       if ['200', '202'].include? response.code
         Unleash.logger.debug "Report sent to unleash server successfully. Server responded with http code #{response.code}"
       else
+        # :nocov:
         Unleash.logger.error "Error when sending report to unleash server. Server responded with http code #{response.code}."
+        # :nocov:
       end
-    end
-
-    private
-
-    def bucket_empty?
-      Unleash.toggle_metrics.features.empty?
     end
   end
 end