+++ lib/unleash/variant.rb
@@ -14,12 +14,18 @@ module Unleash
     end
 
     def to_s
+      # :nocov:
       "<Variant: name=#{self.name},enabled=#{self.enabled},payload=#{self.payload},feature_enabled=#{self.feature_enabled}>"
+      # :nocov:
     end
 
     def ==(other)
       self.name == other.name && self.enabled == other.enabled && self.payload == other.payload \
         && self.feature_enabled == other.feature_enabled
+    end
+
+    def self.disabled_variant
+      Variant.new(name: 'disabled', enabled: false, feature_enabled: false)
     end
   end
 end