+++ spec/unleash/configuration_spec.rb
@@ -203,7 +203,7 @@ RSpec.describe Unleash do
       expect(WebMock).to have_requested(:post, 'http://test-url/api/client/register')
         .with(
           body: hash_including(
-            yggdrasilVersion: nil,
+            yggdrasilVersion: "0.13.2",
             specVersion: anything,
             platformName: anything,
             platformVersion: anything