+++ spec/unleash/metrics_reporter_spec.rb
@@ -8,8 +8,6 @@ RSpec.describe Unleash::MetricsReporter do
     Unleash.logger = Unleash.configuration.logger
     Unleash.logger.level = Unleash.configuration.log_level
     # Unleash.logger.level = Logger::DEBUG
-    Unleash.toggles = []
-    Unleash.toggle_metrics = {}
 
     Unleash.configuration.url         = 'http://test-url/'
     Unleash.configuration.app_name    = 'my-test-app'
@@ -27,24 +25,26 @@ RSpec.describe Unleash::MetricsReporter do
       config.instance_id = 'rspec/test'
       config.disable_client = true
     end
-    Unleash.toggle_metrics = Unleash::Metrics.new
+    Unleash.engine = YggdrasilEngine.new
 
-    Unleash.toggle_metrics.increment('featureA', :yes)
-    Unleash.toggle_metrics.increment('featureA', :yes)
-    Unleash.toggle_metrics.increment('featureA', :yes)
-    Unleash.toggle_metrics.increment('featureA', :no)
-    Unleash.toggle_metrics.increment('featureA', :no)
-    Unleash.toggle_metrics.increment('featureB', :yes)
+    Unleash.engine.count_toggle('featureA', true)
+    Unleash.engine.count_toggle('featureA', true)
+    Unleash.engine.count_toggle('featureA', true)
+    Unleash.engine.count_toggle('featureA', false)
+    Unleash.engine.count_toggle('featureA', false)
+    Unleash.engine.count_toggle('featureB', true)
 
     report = metrics_reporter.generate_report
     expect(report[:bucket][:toggles]).to include(
-      "featureA" => {
+      featureA: {
         no: 2,
-        yes: 3
+        yes: 3,
+        variants: {}
       },
-      "featureB" => {
+      featureB: {
         no: 0,
-        yes: 1
+        yes: 1,
+        variants: {}
       }
     )
 
@@ -74,14 +74,14 @@ RSpec.describe Unleash::MetricsReporter do
       )
       .to_return(status: 200, body: "", headers: {})
 
-    Unleash.toggle_metrics = Unleash::Metrics.new
+    Unleash.engine = YggdrasilEngine.new
 
-    Unleash.toggle_metrics.increment('featureA', :yes)
-    Unleash.toggle_metrics.increment('featureA', :yes)
-    Unleash.toggle_metrics.increment('featureA', :yes)
-    Unleash.toggle_metrics.increment('featureA', :no)
-    Unleash.toggle_metrics.increment('featureA', :no)
-    Unleash.toggle_metrics.increment('featureB', :yes)
+    Unleash.engine.count_toggle('featureA', true)
+    Unleash.engine.count_toggle('featureA', true)
+    Unleash.engine.count_toggle('featureA', true)
+    Unleash.engine.count_toggle('featureA', false)
+    Unleash.engine.count_toggle('featureA', false)
+    Unleash.engine.count_toggle('featureB', true)
 
     metrics_reporter.post
 
@@ -105,7 +105,7 @@ RSpec.describe Unleash::MetricsReporter do
   end
 
   it "does not send a report, if there were no metrics registered/evaluated" do
-    Unleash.toggle_metrics = Unleash::Metrics.new
+    Unleash.engine = YggdrasilEngine.new
 
     metrics_reporter.post
 
@@ -116,15 +116,15 @@ RSpec.describe Unleash::MetricsReporter do
     WebMock.stub_request(:post, "http://test-url/client/metrics")
       .to_return(status: 200, body: "", headers: {})
 
-    Unleash.toggle_metrics = Unleash::Metrics.new
-    Unleash.toggle_metrics.increment('featureA', :yes)
+    Unleash.engine = YggdrasilEngine.new
+    Unleash.engine.count_toggle('featureA', true)
 
     metrics_reporter.post
 
     expect(WebMock).to have_requested(:post, 'http://test-url/client/metrics')
       .with(
         body: hash_including(
-          yggdrasilVersion: nil,
+          yggdrasilVersion: "0.13.2",
           specVersion: anything,
           platformName: anything,
           platformVersion: anything